requirements.txt
google-adk
a2a-sdk
fastmcp
uvicorn
httpx
nest_asyncio
termcolor
python-dotenv
google-generativeai
requests